requirements.txt
alembic==1.10.2
apispec==6.3.0
cachelib==0.9.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-Caching==2.0.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Migrate==4.0.2
Flask-MySQLdb==1.0.1
flask-smorest==0.40.0
Flask-SQLAlchemy==3.0.2
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
bangla==0.0.2
marshmallow-jsonschema==0.13.0
marshmallow-sqlalchemy==0.28.1
mysqlclient==2.1.1
packaging==23.0
passlib==1.7.4
Pillow==9.4.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
python-dotenv==0.21.1
requests==2.28.2
setuptools==65.5.0
six==1.16.0
SQLAlchemy==2.0.0
typing_extensions==4.5.0
urllib3==1.26.15
uuid==1.30
webargs==8.2.0
Werkzeug==2.2.3
numpy==1.24.3
opencv-python==4.7.0.72